fixed lulz mistakes
