+++ src/button.css
@@ -22,8 +22,3 @@
   width: auto;
   margin-right: 20px;
 }
-
-.button-23--small {
-  font-size: 12px;
-  padding: 5px;
-}
